+++ src/main/java/com/thiagoporfirio/elvivo/domain/repositories/user/UserPersonalDataRepository.java
@@ -3,7 +3,7 @@ package com.thiagoporfirio.elvivo.domain.repositories.user;
 import com.thiagoporfirio.elvivo.domain.entities.user.UserPersonalDataEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserPersonalDataRepository extends JpaRepository<UserPersonalDataEntity, Long>
+public interface UserPersonalDataRepository extends JpaRepository<UserPersonalDataEntity, Integer>
 {
     public boolean existsByDni(String dni);
 }
